Migrate nvp-qos to generic name qos-queue
In Juno, the nvp-qos extension was renamed to qos-queue in favor of having
a generic name. In order to maintain backwards compatibility neutron
advertises both nvp-qos and qos-queue (since icehouse). This patch phases out
nvp-qos from nova so that it uses qos-queue so that in Kxxx we'll be able to
drop nvp-qos from neutron completely.

In addition this patch moves qos-queue to a constant like the other extensions
and improves a doc string.

Change-Id: Id43ea931c995865ddc3455abfb625d86dd93e78f
Closes-bug: 1329075
(cherry picked from commit 41bb7dc7b9886a8000c3c1759b2c305ec91824fd)