Avoid resizing disk if the disk size doesn't change
Sometimes we only want to resize the cpu and ram of the instance,
but nova still resizes the disk even if the disk size remains the
same with the original, which costs lots of time.

So we need to avoid resizing disk if the disk size doesn't change.

Change-Id: Ic5f0f21c01c52bec0aa18febc76352bf48df6005
Close-Jira: OPENSTACK-334