libvirt: make snapshot use RBD snapshot/clone when available
When rbd is used for ephemeral disks and image storage, make snapshot
use ceph directly, and update glance with the new location.  In
case of failure, it will gracefully fallback to the "generic"
snapshot method.
This requires changing the typical permissions for the nova ceph
user (if using authx) to allow writing to the pool where vm images
are stored.

Change-Id: I97e7f7c7a8169c908e916385d35bf744e6e3fceb